Removed some warnings
git-svn-id: https://svn.code.sf.net/p/wired/code/trunk/wired@394 6f5cd6f6-390f-0410-abd1-b47fe60d5253

+++ src/dssi/WiredExternalPluginGui.cpp
@@ -108,7 +108,7 @@ wxWindow	*WiredDSSIGui::CreateView(wxWindow *rack, wxPoint &pos, wxSize &size)
       FaderIndex[i + 1] = iter->first;
       cout << "** " << iter->second.Data.LowerBound << "<" << *(iter->second.Data.Data) << "<" 
 	   << iter->second.Data.UpperBound << endl;
-      Faders[i]->SetValue(*(iter->second.Data.Data) / (iter->second.Data.UpperBound - iter->second.Data.LowerBound) * 100);
+      Faders[i]->SetValue((int)*(iter->second.Data.Data) / (iter->second.Data.UpperBound - iter->second.Data.LowerBound) * 100);
       // gruik bis
       if (i == 13)
 	break;
